feat(tests): add filesystem error scenario tests and logging fixes
- Add unit tests for disk space exhaustion and permission denied scenarios.
- Include test coverage for file creation, write, and access operations under error conditions.
- Update `release_action_plan.md` to mark filesystem error scenario tests as complete.
- Fix `LogMethodExit` error type casting to align with expectations.
- Add null parent handling in `inode.go` to prevent potential panics.

Refs: #78, #85

+++ internal/fs/file_operations.go
@@ -409,7 +409,7 @@ func (f *Filesystem) Write(_ <-chan struct{}, in *fuse.WriteIn, data []byte) (ui
 	id := f.TranslateID(in.NodeId)
 	inode := f.GetID(id)
 	if inode == nil {
-		defer logging.LogMethodExit(methodName, time.Since(startTime), uint32(0), fuse.EBADF)
+		defer logging.LogMethodExit(methodName, time.Since(startTime), uint32(0), int32(fuse.EBADF))
 		return 0, fuse.EBADF
 	}
 
@@ -443,7 +443,7 @@ func (f *Filesystem) Write(_ <-chan struct{}, in *fuse.WriteIn, data []byte) (ui
 			logging.FieldOperation, "file_write",
 			logging.FieldID, id,
 			logging.FieldPath, path)
-		defer logging.LogMethodExit(methodName, time.Since(startTime), uint32(0), fuse.EIO)
+		defer logging.LogMethodExit(methodName, time.Since(startTime), uint32(0), int32(fuse.EIO))
 		return 0, fuse.EIO
 	}
 
@@ -457,7 +457,7 @@ func (f *Filesystem) Write(_ <-chan struct{}, in *fuse.WriteIn, data []byte) (ui
 			logging.FieldPath, path,
 			logging.FieldOffset, offset,
 			logging.FieldSize, nWrite)
-		defer logging.LogMethodExit(methodName, time.Since(startTime), uint32(n), fuse.EIO)
+		defer logging.LogMethodExit(methodName, time.Since(startTime), uint32(n), int32(fuse.EIO))
 		return uint32(n), fuse.EIO
 	}
 
@@ -479,7 +479,7 @@ func (f *Filesystem) Write(_ <-chan struct{}, in *fuse.WriteIn, data []byte) (ui
 			Msg("File write completed successfully")
 	}
 
-	defer logging.LogMethodExit(methodName, time.Since(startTime), uint32(n), fuse.OK)
+	defer logging.LogMethodExit(methodName, time.Since(startTime), uint32(n), int32(fuse.OK))
 	return uint32(n), fuse.OK
 }
